Fix build issues and update
Not sure what was going on. ESLint was throwing a bunch of errors about
'any' types returned by the prisma client. Ran npm doctor and fixed
outstanding issues there. Ran a full upgrade of packages with `npm
update` and then regenerated the prisma client. Fixed how we were
destructuring data in the call that loads the practice session, and
added a couple of missing async annotations in the routers.